+++ src/bract/cli/internal.clj
@@ -10,8 +10,8 @@
 (ns bract.cli.internal
   (:require
     [clojure.string    :as string]
-    [bract.cli.config  :as clim-config]
-    [bract.core.config :as core-config]
+    [bract.cli.keydef  :as clim-kdef]
+    [bract.core.keydef :as core-kdef]
     [bract.core.util   :as core-util]))
 
 
@@ -28,11 +28,11 @@
     context
     (let [verbose (get-in parse-result [:options :verbose])]
       (cond
-        (true? verbose)  (assoc context (key core-config/ctx-verbose?) true)
-        (false? verbose) (assoc context (key core-config/ctx-verbose?) false)
-        (nil? verbose)   (if (contains? context (key core-config/ctx-verbose?))
+        (true? verbose)  (assoc context (key core-kdef/ctx-verbose?) true)
+        (false? verbose) (assoc context (key core-kdef/ctx-verbose?) false)
+        (nil? verbose)   (if (contains? context (key core-kdef/ctx-verbose?))
                            context
-                           (assoc context (key core-config/ctx-verbose?) false))
+                           (assoc context (key core-kdef/ctx-verbose?) false))
         :otherwise       context))))
 
 
@@ -44,8 +44,8 @@
       (as-> config-filenames <>
         (string/split <> #",")
         (mapv string/trim <>)
-        (assoc context (key core-config/ctx-config-files) <>))
-      (if (clim-config/ctx-config-required? context)
+        (assoc context (key core-kdef/ctx-config-files) <>))
+      (if (clim-kdef/ctx-config-required? context)
         (do
           (core-util/err-println "No config file specified as argument")
           (core-util/err-println (get-in parse-result [:summary]))
@@ -58,7 +58,7 @@
   (if (reduced? context)
     context
     (let [command (get-in parse-result [:options :command] "run")]
-      (assoc context (key clim-config/ctx-command) command))))
+      (assoc context (key clim-kdef/ctx-command) command))))
 
 
 (defn assoc-cmd-args
@@ -66,4 +66,4 @@
   (if (reduced? context)
     context
     (let [arguments (get-in parse-result [:arguments] [])]
-      (assoc context (key clim-config/ctx-cmd-args) arguments))))
+      (assoc context (key clim-kdef/ctx-cmd-args) arguments))))